updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -38,8 +38,9 @@ notes = {1: {'title': 'First Note', 'text': 'This is my first note', 'date': '10
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='mstout12@uncc.edu').one()
-    return render_template('index.html', user=a_user)
+    if session.get('user'):
+        return render_template('index.html', user=session['user'])
+    return render_template("index.html")
 
 @app.route('/notes')
 def get_notes():
@@ -59,49 +60,56 @@ def get_note(note_id):
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
     #a_user = {'name': 'Brandon', 'email': 'bran@uncc.edu'}
-    print("Request method used is: ", request.method)
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        from datetime import date
-        today = date.today()
-        today = today.strftime("%m-%d-%y")
-        new_entry = Note(title, text, today)
-        db.session.add(new_entry)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+    #print("Request method used is: ", request.method)
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            from datetime import date
+            today = date.today()
+            today = today.strftime("%m-%d-%y")
+            new_entry = Note(title, text, today)
+            db.session.add(new_entry)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='mstout12@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    #check method used for request
-    if request.method == 'POST':
-        #get Title data
-        title = request.form['title']
-        #get text data
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
-        #update note data
-        note.title = title
-        note.text = text
-        #update note in db
-        db.session.add(note)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+    if session.get('user'):
+        #check method used for request
+        if request.method == 'POST':
+            #get Title data
+            title = request.form['title']
+            #get text data
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
+            #update note data
+            note.title = title
+            note.text = text
+            #update note in db
+            db.session.add(note)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+            return render_template('new.html', note=my_note, user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='mstout12@uncc.edu').one()
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-        return render_template('new.html', note=my_note, user=a_user)
+        return redirect(url_for('login'))
 
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
-    #retrieve note from db
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-    return redirect(url_for('get_notes'))
+    if session.get('user'):
+        #retrieve note from db
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect(url_for('login'))
 
 @app.route('/register', methods=['POST', 'GET'])
 def register():
